// typescript/interfaces.ts
export interface IReservationQueryParams {
  page?: number;
  limit?: number;
  sort?: {
    order?: "asc" | "desc";
    field?: string;
  };
  search?: string;
  status?: string[]; // Array of strings, as multiple statuses can be included
  vendor_type?: string[]; // Array of strings for vendor types
  category_id?: string[]; // Assuming this is an array of strings, given the context
  sub_category_id?: string[]; // Also an array of strings for sub-category IDs
  reservation_date?: {
    start_date?: string; // Format: YYYY-MM-DD
    end_date?: string; // Format: YYYY-MM-DD
  };
  address?: string;
}

export interface ReservationResponse {
  data: Reservation[];
  limit: number;
  message: string;
  page: number;
  pages: number;
  success: boolean;
  total: number;
}

export interface Reservation {
  _id: string;
  address: string;
  category_id: string;
  category_title: string;
  client_data: ClientData;
  date: string;
  gift_wrapper: boolean;
  isDeleted: boolean;
  note: string;
  package_price: number;
  package_title: string;
  ref_no: string;
  service_data: ServiceData;
  status: string;
  //   time: string;
  start_time: string;
  end_time: string;
  user_id: string;
  vendor_data: VendorData;
  createdAt: string;
}

interface ClientData {
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: "individual" | "company" | string;
  uid: string;
  _id: string;
}

interface ServiceData {
  address: string;
  city: string;
  country: string;
  description: string;
  images: string[];
  state: string;
  title: string;
  zip_code: string;
}

interface VendorData {
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: "individual" | "company" | string;
  uid: string;
  _id: string;
}

export interface CategoryResponse {
  success: boolean;
  data: Category[];
  message: string;
}

interface Category {
  _id: string;
  name: string;
  sub_categories: SubCategory[];
  total_services: number;
  order_num: number;
  isDeleted: boolean;
  status: "Active" | string;
}
interface SubCategory {
  name: string;
  _id: string;
  total_services: number;
}

// export interface ReservationDetailsResponse {
//   data: BookingDetails;
//   message: string;
//   success: boolean;
// }
export interface ReservationInfoResponse {
  success: boolean;
  message: string;
  data: ReservationData;
}

interface ReservationData {
  _id: string;
  user_id: string;
  service_title: string;
  service_description: string;
  category_title: string;
  category_id: string;
  sub_category_id: string;
  package_title: string;
  package_price: number;
  package_description: string;
  service_data: ServiceData;
  client_data: ClientData;
  vendor_data: VendorData;
  ref_no: string;
  address: string;
  date: string;
  start_time: string;
  end_time: string;
  gift_wrapper: boolean;
  note: string;
  isDeleted: boolean;
  status: string;
}
// ...
interface CancellationPolicy {
  type: string;
  description: string;
}
// ...
export interface ServiceFetchResponse {
  success: boolean;
  message: string;
  data: ServiceDataTwo[];
  limit: number;
  page: number;
  pages: number;
  total: number;
  grouped_counts: GroupedCounts;
}

export interface ServiceDataTwo {
  _id: string;
  user_id: string;
  vendor_data: VendorData;
  category_title: string;
  packages_data: PackageData[];
  title: string;
  packages: string[];
  description: string;
  category: string;
  sub_category: string;
  images: string[];
  address: string;
  country: string;
  city: string;
  state: string;
  zip_code: string;
  excluded_areas: string[];
  operating_area: string;
  booking_type: string;
  rules: any[];
  cancellation_policy: CancellationPolicy;
  currency: string;
  createdAt: string;
  updatedAt: string;
  status: string;
  isDeleted: boolean;
}

interface VendorData {
  _id: string;
  uid: string;
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: string;
}

interface PackageData {
  _id: string;
  name: string;
  description: string;
  duration: string;
  features: string[];
  rate: number;
}

interface CancellationPolicy {
  type: string;
  description: string;
}

interface GroupedCounts {
  all: number;
  company: number;
  individual: number;
}